Moved env preset configuration and debug info into web-worker This isn't quite working yet but I'm going to commit to share with others.
